Simplify connector and examples
